Style: Clarify deviations from the GCSG
Corresponding configurations for clang-format are provided.
These deviations are tentative
as observed from the current state of the codebase.

Note that the observed deviations are not
from the core, contributing developer or author,
so many deviations or implicit conventions may be missing.
It is best to get this information
from people who actually wrote the code.

This commit addresses issue #238.

[skip ci]