AddressBook: fix two publisher/subscription-related bugs
Noticed upon reviewing CID 151160/2/6, control was never allowed to
reach the loading of publishers if a subscription was already loaded at
startup - thus publishers were never loaded.

In addition to publishers not being loaded, the subscriber
would always be in a "downloading" state (though it was never
downloading) - thus allowing the thread to sleep unnecessarily upon
shutdown.

Fixes #445 #446

References #305 #337 #263