SSUSession: trivial patch to handle segfault in #119
A trivial patch until the bigger issue is resolved; this will prevent
sending SessionCreated to the very few routers who send bad?
SessionRequest data (at least uninitialized Diffie-Hellman?).
It's not ideal but it works for now, and is better than what we have at
the moment (which is nothing).

Note: this is not related to ParseHeader() which, despite the TODO, even
after initializing m_Data for the length of keying material before
moving pointer with ConsumeData(), the segfault will occur during key
creation (without this patch).