Crypto: move exception handling into core/client
As noted in the TODO's, most of this refactor is immediately ineffective
if the exception isn't handled at a higher calling level *but*, when/if
things are thrown and *not* caught, this refactor should get us started
in the right direction in terms of debugging and subsequent higher-level
re-design.

As I noted in yesterday's meeting:

"JFTR: of the thousands of lines of code that we forked from, almost
none of it handled exceptions. All this spaghetti code, and that guy
did not even have a basic grasp of design with exception handling in
mind. Literally, much of the core code is still stringed together with
the hope that it won't fail. I hate it. It's always been like walking
on glass, but a real fix will take time. So, I've started in that
direction at least in terms of getting exceptions handled or handled
better."

Also note the TODO's for the exception class: this work is a WIP so I do
not want to refactor the remaining code-base with this class until it's
solidified.