Build: fix GCC 7 implicit fallthrough warnings
Note: GCC 7 will also recognize the [[clang::fallthrough]] attribute but
we currently avoid doing this because of the older compilers we use.

From https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html:

>C++17 provides a standard way to suppress the -Wimplicit-fallthrough
warning using [[fallthrough]]; instead of the GNU attribute. In C++11 or
C++14 users can use [[gnu::fallthrough]];, which is a GNU extension.
Instead of these attributes, it is also possible to add a fallthrough
comment to silence the warning.