Boost.Python: implement core/client wrappers + utils
Includes tweaking the client API.

Although we may ultimately wish for the client to maintain control
of a core instance (as it currently does, once both core/client are
initialized), enforcing this behaviour with unique_ptr may be an
anti-pattern and it seems that we'll want to sacrifice move semantics
anyway for a non-const reference approach to mutable core members.

WIP. See TODOs.