Proxy: Catch a potential exception in destructor.
shared_from_this() can throw std::bad_weak_ptr if called on an
object that has never been shared in a shared_ptr. Throwing an exception in
a destructor crashes the program.

Not using HTTPProxyHandler with shared pointers is a serious bug and
the program probably should crash, but in a more controlled manner and
with an error message.

But in mainly silences Coverity #175240.

Handling the exception using the exception dispatcher. See PR #923 .

Signed-off-by: Tadeas Moravec <moravec.tadeas@gmail.com>