Socks: Catch a potential exception in destructor.
shared_from_this() can throw std::bad_weak_ptr if called on an
object that has never been shared in a shared_ptr. Throwing an exception
in a destructor crashes the program.

Not using SOCKSHandler with shared pointers is a serious bug and
the program probably should crash, but in a more controlled manner and
with an error message.

But in mainly silences Coverity #175240.

Same situation as in PR #923 . I just haven't noticed that while working on
that one.

Signed-off-by: Tadeas Moravec <moravec.tadeas@gmail.com>