CMake: side-step Clang (CMake?) bug for Release build
This is a very strange bug. See the diff and #993 for details.

The only difference between our Release flags and Debug flags are the
defining of NDEBUG. As we are in Alpha, this is not a conern but this
patch must be removed no later than a Beta (preferably Alpha) release.